requirements.txt
langchain-community
langchain
ipykernel
python-dotenv
langchain_core
streamlit
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence-transformers
langchain-huggingface
faiss-cpu
duckdb
pandas
langchain_groq
langchain_chroma
ddgs==0.6.2